Feature/footer menu tokenization (#3)
* Refactor: Remove hardcoded values from appz_footer_menu

Removed hardcoded values from the `appz_footer_menu.dart` file and replaced them with tokens from `supporting_tokens.json`. This improves the maintainability and scalability of the component.

* Refactor: Remove hardcoded values from appz_footer_menu

Removed hardcoded values from the `appz_footer_menu.dart` file and replaced them with tokens from `supporting_tokens.json`. This improves the maintainability and scalability of the component.

* Fix: Footer menu rendering on mobile screens

I refactored the `appz_footer_menu.dart` to use a `BottomAppBar` and a custom `NavClipper` to fix the rendering issue on mobile screens. This ensures that the footer menu is displayed correctly across all screen sizes.

* Fix: Footer menu rendering on mobile screens

I adjusted the `notchMargin` and `gapLocation` properties of the `AnimatedBottomNavigationBar` to fix the rendering issue on mobile screens. This ensures that the footer menu is displayed correctly across all screen sizes.

* Fixed the right corner issue.

---------
Authored-by: Syed Tajamul Shah